api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
